requirements.txt
torch>=2.5,<3.0
loguru>=0.7.3,<0.8.0
pytest>=8.3.4,<9.0.0
jupyter>=1.1.1,<2.0.0
pymilvus>=2.5.0,<3.0.0
open_clip_torch>=2.29.0,<3.0.0
fiftyone>=1.1.0,<2.0.0
opencv-python>=4.10.0.84,<5.0.0
gradio>=5.8.0,<6.0.0
ruff>=0.8.2,<0.9.0